Move map utils from index to utils.ts
